♻️ D97M2G task: decompose observations command specs
