+++ .agentplane/tasks/202605290416-D97M2G/pr/diffstat.txt
@@ -1,0 +1,3 @@
+ .../src/commands/task/observations.command.ts      | 249 ++-------------------
+ .../src/commands/task/observations.specs.ts        | 230 +++++++++++++++++++
+ 2 files changed, 252 insertions(+), 227 deletions(-)
