+++ .agentplane/tasks/202605290416-D97M2G/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 D97M2G task: Task observations command decomposition [202605290416-D97M2G]
