+++ .agentplane/tasks/202605290416-D97M2G/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290416-D97M2G/task-observations-command-decomposition",
+  "created_at": "2026-05-29T04:16:21.019Z",
+  "diffstat_sha256": "sha256:8878237599aca0ea85988326cba77c024b0a41ec8498c7206610e508a0b5a7af",
+  "last_verified_at": "2026-05-29T04:21:38.744Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290416-D97M2G",
+  "updated_at": "2026-05-29T04:16:21.019Z",
+  "verify": {
+    "status": "pass"
+  }
+}
