+++ .agentplane/tasks/202605290416-D97M2G/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T04:16:21.019Z
+
+## Task
+
+- Task: `202605290416-D97M2G`
+- Title: Task observations command decomposition
+- Status: DOING
+- Branch: `task/202605290416-D97M2G/task-observations-command-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290416-D97M2G/README.md`
+
+## Verification
+
+- State: ok
+- Note: Observed: observations command specs were extracted into observations.specs.ts; observations.command.ts is below hotspot threshold. Checks: observations.unit.test, typecheck, arch:check, knip:check, lint:core, format:changed, hotspots:check passed.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T04:16:21.019Z
+- Branch: task/202605290416-D97M2G/task-observations-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/commands/task/observations.command.ts      | 249 ++-------------------
+ .../src/commands/task/observations.specs.ts        | 230 +++++++++++++++++++
+ 2 files changed, 252 insertions(+), 227 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
